+++ src/main/java/com/jootcamp/superboard/common/constants/UserConstant.java
@@ -2,5 +2,5 @@ package com.jootcamp.superboard.common.constants;
 
 public class UserConstant {
     // todo : 로그인 구현 이후 삭제 예정
-    public static final long USER_ID = 2024L;
+    public static final String USER_INFO ="userInfo";
 }
